Bug Fixes & Search Enhancements
Fixed statblocks not loading correctly if they did not contain a legendary action or special action array. Changed searching to load on enter instead of on key press. Added search tags to display CR and monster type, which will be further enhanced in a future update to allow selactable and robust search filtering criteria.